# Swagger Documentation.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Article/GetById": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/getByTheme": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/getByName": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/getByLength": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/getByRangeOfRate": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "max",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "min",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/AddAricle": {
      "post": {
        "tags": [
          "Article"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/UpdateArticle": {
      "put": {
        "tags": [
          "Article"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleAddmodel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Article/delete": {
      "delete": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ArticleRate": {
      "get": {
        "tags": [
          "ArticleRate"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ArticleRate"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleRateAddModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleRateAddModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleRateAddModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleRateAddModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ArticleRate/delete": {
      "delete": {
        "tags": [
          "ArticleRate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "authorFullName": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "themeId": {
            "type": "integer",
            "format": "int32"
          },
          "theme": {
            "$ref": "#/components/schemas/Theme"
          },
          "articleRates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleRate"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ArticleAddmodel": {
        "required": [
          "authorFullName",
          "body",
          "themeName",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 70,
            "minLength": 5,
            "type": "string"
          },
          "body": {
            "maxLength": 5000,
            "minLength": 50,
            "type": "string"
          },
          "authorFullName": {
            "maxLength": 100,
            "type": "string"
          },
          "themeName": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string"
          },
          "theme": {
            "$ref": "#/components/schemas/Theme"
          }
        },
        "additionalProperties": false
      },
      "ArticleRate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rate": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "article": {
            "$ref": "#/components/schemas/Article"
          }
        },
        "additionalProperties": false
      },
      "ArticleRateAddModel": {
        "required": [
          "articleTitle",
          "rate",
          "userLogin"
        ],
        "type": "object",
        "properties": {
          "rate": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "maxLength": 100,
            "minLength": 10,
            "type": "string",
            "nullable": true
          },
          "articleTitle": {
            "maxLength": 70,
            "minLength": 5,
            "type": "string"
          },
          "userLogin": {
            "maxLength": 20,
            "minLength": 4,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Theme": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Article"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}